Tasks/requirements.txt
pandas==1.2.4
python-binance==0.7.11